Unit tests!
As I work through the 'nest :: Int -> Doc -> Doc'
problem in Chapter 5 of Real World Haskell I realized
I'd like to solve the problem naively and then in a
way that took into account whether or not parsing was
"in" a string. Those are both things I'd like to have
as test cases so that as I solve the first I can see
how I do with the second and as I solve the second
I can prove no regressions of the first case.

What I learned is that I understand how to find library
names on hackage, how to add them to the cabal file,
how to read online documentation for just what I need
and how to get things to build through VS Code (running
ghci on the main.hs). I don't yet know how to create a
whole 'test' branch of my cabal file though so the
tests are still main.hs but Bond does this so I should
be able to figure that out.

I also learned that Tasty is one level of abstraction
higher (of course) than a type of test runner and is
a runner of runners. So rather than *being* HUnit Tasty
runs HUnit, and QuickCheck or whatever. I also may have
gotten a peek at what property testing is: rather than
tests at the example level, it's tests at the class of
example level. Still working on that though.