Either Bind allows type of Right to change
Previously two Eithers bound through Bind would be required to have the same type for Right. We relax this restriction so that now one can change the type within Right but still use Bind to create serial operations on Either. The use-case for this is if you are running multiple operations that could fail but are building up information as you go which is represented in a series of different (perhaps more detailed) types.