Folding example from Chapter 10
The key key key thing to note about how foldr is able to
recurse into a bracketed structure is that the folding
function is only partially applied. The folding function
has the left argument (single x) partially applied while
foldr must unwrap any remaining list items in concordance
to the wishes of however strict the folding function is
in order to satisfy the arguments of the folding
function with respect to being able to get to weak head
normal form (aka. evaluate fully).