Handle non-letters and capitals, new functions
In addition to the mostFrequentLetter functions, in order
to answer the question of "which sentence has the most of
any particular letter" I created mostFrequentLettterAndMagnitude
functions which preserve the magnitude of the most frequent
letter for a given sentence so I can compare that with other
sentences.

I have also learned that using !! is terrible and should not
be done.