Figure out problem with function
I had thought that creating a 'case of' expression would
result in me being able to create a name that would shadow
the same name in the larger context. Not only does this
not appear to be true, but it's just sloppy. Also I was
just not using the right values in my enclosed function.