First ever Haskell https call!
Figured out how to make an https call which will be required for the next Project Rosalind problem which involves pulling fasta strings from a website.

This was an accomplishment because I couldn't use Network.HTTP.Client.Simple and had to figure out how to pass in the TLS manager myself.