Figured out how to read a local file
I can't read a file not controlled by the Haskell app, however that's defined. I *can* read a file that was created by and within the Haskell app. I have no idea what the difference between a file owned by the app and one sitting on the Desktop but anyway, this works.