FindAllPossibleSubstrings with index
So the working theory is that to find all possible matches in a string we need to find all overlapping matches. To do this we need to find all possible substrings and then run our matcher over them. This allows us to use a regex which could otherwise not find overlapping matches. It also allows us to do that without writing another Parsec parser.

So this commit brings us to a compiling, working, quickChecking findAllSubstrings that along with producing the substrings also records the exact index in which they were discovered so we can later differentiate strings found that are otherwise identical according to the regex.

The plan is to run this matcher side by side with our non-overlapping matcher to prove we're finding >= matches.