Pay-dirt: list comprehensions to generate slices
The first substring will be the entire string. Subsequently we'll build all possible substrings by generating all possible end-points of each of those substrings. By generating our endpoints in such a way that the previous string contains the extent of the substring we are generating this loop (but with one extra character) we should be able to create the listArray with a function guided by the comprehension to create all the slices and thus generate each entry in the array without recursion.