Start playing around with vectors
I'm learning that up until now I had no "approach" to learning a module. If some page turned me on to a function in a module I'd just take that one function. But lately I found myelf struggling for more power and flexibility and realized that the solution is to read the module docs top to bottom with focus on which functions there are, what scenarios they cover and what their performance is.

So now rather than List I'm hoping to use Vector because it has a lot of O(1) operations I want, it's super easy to go back and forth between List and Vector and the module has a lot of little functions I might need and they operate at O(1).